uint16 -> uint64
# message auto-generated for no-merge-commit merge:
merge linux into master

hccl calculator GroupData 数据类型从uint16变更为uint64

Created-by: msprofDeveloper

Match-id-e28c5cd63f3e907c7870cf17d457c7627df40eb5

+++ analysis/csrc/domain/services/association/calculator/hccl/hccl_calculator.cpp
@@ -27,7 +27,7 @@ namespace Domain {
 namespace {
 struct GroupData {
     uint64_t firstTimestamp = UINT64_MAX;
-    uint16_t count = 0;
+    uint64_t count = 0;
 };
 
 struct OpTypeInfo {
